Fixed 2 validation errors & 1 warning. + SEO
Removed one validation warning because of role="navigation"
and two errors because of <meta> tags inside ul. 
This way it completely validates at https://validator.nu/ https://validator.w3.org/nu/ and https://developers.google.com/structured-data/testing-tool/
Above tests have been made to pages in 3 level nested Breadcrumbs

Also changed h2 to p for SEO reasons because there is no need on any website to emphasize at the content of show_browse with an h2 tag Headings should be used only for content that is important and related to the topic of the current page.